+++ app/src/main/java/com/concordium/wallet/ui/bakerdelegation/baker/BakerRegistrationCloseActivity.kt
@@ -112,12 +112,6 @@ class BakerRegistrationCloseActivity :
         startActivityForResultAndHistoryCheck(intent)
     }
 
-    override fun transactionSuccessLiveData() {
-    }
-
     override fun errorLiveData(value: Int) {
-    }
-
-    override fun showDetailedLiveData(value: Boolean) {
     }
 }
