+++ app/src/main/java/com/concordium/wallet/ui/bakerdelegation/baker/BakerRegistrationOpenActivity.kt
@@ -56,12 +56,6 @@ class BakerRegistrationOpenActivity :
         startActivityForResultAndHistoryCheck(intent)
     }
 
-    override fun transactionSuccessLiveData() {
-    }
-
     override fun errorLiveData(value: Int) {
-    }
-
-    override fun showDetailedLiveData(value: Boolean) {
     }
 }
