+++ app/src/main/java/com/concordium/wallet/ui/bakerdelegation/delegation/DelegationStatusActivity.kt
@@ -16,8 +16,6 @@ import com.concordium.wallet.ui.bakerdelegation.common.StatusActivity
 import com.concordium.wallet.ui.bakerdelegation.delegation.introflow.DelegationRemoveIntroFlowActivity
 import com.concordium.wallet.ui.bakerdelegation.delegation.introflow.DelegationUpdateIntroFlowActivity
 import com.concordium.wallet.ui.common.GenericFlowActivity
-import com.concordium.wallet.util.DateTimeUtil.formatTo
-import com.concordium.wallet.util.DateTimeUtil.toDate
 import kotlinx.android.synthetic.main.delegationbaker_status.*
 
 class DelegationStatusActivity :
@@ -34,74 +32,52 @@ class DelegationStatusActivity :
             findViewById<ImageView>(R.id.status_icon).setImageResource(R.drawable.ic_logo_icon_pending)
             setContentTitle(R.string.delegation_status_content_empty_title)
             setEmptyState(getString(R.string.delegation_status_content_empty_desc))
-
             status_button_bottom.text = getString(R.string.delegation_status_continue)
             status_button_bottom.setOnClickListener {
                 continueToCreate()
             }
+            return
         }
-        else {
-            findViewById<ImageView>(R.id.status_icon).setImageResource(R.drawable.ic_big_logo_ok)
-            setContentTitle(R.string.delegation_status_content_registered_title)
-            addContent(R.string.delegation_status_content_delegating_account, account.name + "\n\n" + account.address)
-            addContent(R.string.delegation_status_content_delegation_amount, CurrencyUtil.formatGTU(accountDelegation.stakedAmount, true))
-            if (accountDelegation.delegationTarget.delegateType == DelegationTarget.TYPE_DELEGATE_TO_BAKER) {
-                addContent(R.string.delegation_status_content_target_pool, accountDelegation.delegationTarget.bakerId.toString())
-            }
-            else {
-                addContent(R.string.delegation_status_content_target_pool, getString(R.string.delegation_register_delegation_passive_long))
-            }
 
-            if (accountDelegation.restakeEarnings) addContent(R.string.delegation_status_content_rewards_will_be, getString(R.string.delegation_status_added_to_delegation_amount))
-            else addContent(R.string.delegation_status_content_rewards_will_be, getString(R.string.delegation_status_at_disposal))
+        status_button_bottom.text = getString(R.string.delegation_status_update)
 
-            viewModel.bakerDelegationData.account?.accountDelegation?.pendingChange?.let { pendingChange ->
-                val prefix = pendingChange.effectiveTime.toDate()?.formatTo("yyyy-MM-dd")
-                val postfix = pendingChange.effectiveTime.toDate()?.formatTo("HH:mm")
-                val dateStr = getString(R.string.delegation_status_effective_time, prefix, postfix)
-                addContent(getString(R.string.delegation_status_content_take_effect_on) + "\n" + dateStr, "")
-                if (pendingChange.change == "RemoveStake") {
-                    status_button_top.isEnabled = false
-                    addContent(getString(R.string.delegation_status_content_delegation_will_be_stopped), "")
-                } else if (pendingChange.change == "ReduceStake") {
-                    pendingChange.newStake?.let { newStake ->
-                        addContent(getString(R.string.delegation_status_new_amount), CurrencyUtil.formatGTU(newStake, true))
-                    }
-                }
-            }
+        if (viewModel.bakerDelegationData.isTransactionInProgress) {
+            addWaitingForTransaction(R.string.delegation_status_waiting_to_finalize_title, R.string.delegation_status_waiting_to_finalize)
+            return
+        }
 
-            status_button_top.visibility = View.VISIBLE
-            status_button_top.text = getString(R.string.delegation_status_stop)
-            status_button_top.setOnClickListener {
-                continueToDelete()
-            }
+        findViewById<ImageView>(R.id.status_icon).setImageResource(R.drawable.ic_big_logo_ok)
+        setContentTitle(R.string.delegation_status_content_registered_title)
 
-            // no pending changes (either empty or equal to NoChange) -> enabled
-            // accBalance -> if you have pending changes -> disabled
-            viewModel.bakerDelegationData.account?.accountDelegation?.pendingChange.let { pendingChange ->
-                status_button_top.isEnabled = pendingChange == null  || pendingChange.change == PendingChange.CHANGE_NO_CHANGE
-            }
+        addContent(R.string.delegation_status_content_delegating_account, account.name + "\n\n" + account.address)
+        addContent(R.string.delegation_status_content_delegation_amount, CurrencyUtil.formatGTU(accountDelegation.stakedAmount, true))
 
-            // bakerPool -> bakerstakependingchange == RemovePool (aka the baker closed the pool) -> enabled
-            viewModel.bakerDelegationData.bakerPoolStatus?.bakerStakePendingChange?.pendingChangeType.let { pendingChangeType ->
-                if (pendingChangeType == BakerStakePendingChange.CHANGE_REMOVE_POOL) {
-                    status_button_top.isEnabled = true
-                }
-            }
+        if (accountDelegation.delegationTarget.delegateType == DelegationTarget.TYPE_DELEGATE_TO_BAKER) addContent(R.string.delegation_status_content_target_pool, accountDelegation.delegationTarget.bakerId.toString())
+        else addContent(R.string.delegation_status_content_target_pool, getString(R.string.delegation_register_delegation_passive_long))
 
-            status_button_bottom.text = getString(R.string.delegation_status_update)
-            status_button_bottom.setOnClickListener {
-                continueToUpdate()
+        if (accountDelegation.restakeEarnings) addContent(R.string.delegation_status_content_rewards_will_be, getString(R.string.delegation_status_added_to_delegation_amount))
+        else addContent(R.string.delegation_status_content_rewards_will_be, getString(R.string.delegation_status_at_disposal))
+
+        viewModel.bakerDelegationData.account?.accountDelegation?.pendingChange?.let { pendingChange ->
+            addPendingChange(pendingChange, R.string.delegation_status_effective_time, R.string.delegation_status_content_take_effect_on, R.string.delegation_status_content_delegation_will_be_stopped, R.string.delegation_status_new_amount)
+            status_button_top.isEnabled = pendingChange.change == PendingChange.CHANGE_NO_CHANGE
+        }
+
+        status_button_top.visibility = View.VISIBLE
+        status_button_top.text = getString(R.string.delegation_status_stop)
+
+        status_button_top.setOnClickListener {
+            continueToDelete()
+        }
+
+        viewModel.bakerDelegationData.bakerPoolStatus?.bakerStakePendingChange?.pendingChangeType.let { pendingChangeType ->
+            if (pendingChangeType == BakerStakePendingChange.CHANGE_REMOVE_POOL) {
+                status_button_top.isEnabled = true
             }
         }
 
-        //If there are transactions in progress
-        if(viewModel.bakerDelegationData.isTransactionInProgress){
-            status_button_top.isEnabled = false
-            status_button_bottom.isEnabled = false
-            clearState()
-            setContentTitle(R.string.delegation_status_waiting_to_finalize_title)
-            setEmptyState(getString(R.string.delegation_status_waiting_to_finalize))
+        status_button_bottom.setOnClickListener {
+            continueToUpdate()
         }
     }
